+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ Added
 ~~~~~
 - ``conkit.plot`` subpackage refactored to allow ``matplotlib`` access of ``Figure`` instances. This provides
   functionality similar to ``seaborn``, so ``matplotlib.Axes`` can be provided into which a plot is drawn.
+- ``ContactMap.as_list`` function to represent the contact map as a 2D-list of residue indexes
 Changed
 ~~~~~~~
 - Changed API interface for ``conkit.plot`` in accordance to necessary changes for above